requirements.txt
-e .
youtube-transcript-api
langchain-community
langchain-openai
langchain-text-splitters
streamlit
watchdog
pytest
peewee
python-dotenv
chromadb
langchain-chroma
randomname
tiktoken
openai-whisper
pytube